pyproject.toml
[project]
name = "tfm-data"
version = "0.1.0"
description = "Terraforming Mars data analysis"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "openpyxl>=3.0.0",
    "matplotlib>=3.10.8",
]

[project.scripts]
preprocess = "preprocess:run_preprocessing"

[tool.uv]
dev-dependencies = []
